Design: design joinForm
