+++ src/App.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Login from "./pages/Login";
 import Mainpage from "./pages/Mainpage";
 import Mypage from "./pages/Mypage"
 import Home from "./pages/Home";
-import Join from "./pages/Join";
+import JoinForm from "./pages/JoinForm";
+import LoginForm from "./pages/LoginForm";
 
 const App = () => {
   return (
@@ -12,8 +12,8 @@ const App = () => {
       <Routes>
         <Route path="/" element={<Home />} />
 
-        <Route path="/login" element={<Login />} />
-        <Route path="/join" element={<Join />} />
+        <Route path="/login" element={<LoginForm />} />
+        <Route path="/join" element={<JoinForm />} />
         <Route path="/mainpage" element={<Mainpage />} />
         <Route path="/mypage" element = {<Mypage />} />
