+++ src/pages/JoinForm.js
@@ -1,0 +1,69 @@
+import React, { useState } from 'react';
+import api from '../api';
+import { useNavigate } from 'react-router-dom';
+import '../style/AuthForm.css'; 
+
+const Join = () => {
+  const [formData, setFormData] = useState({
+    username: '',
+    password: '',
+    passwordCheck: '',
+    email: '',
+    name: ''
+  });
+
+  const [error, setError] = useState('');
+  const navigate = useNavigate();
+
+  const handleChange = (e) => {
+    setFormData((prev) => ({
+      ...prev,
+      [e.target.name]: e.target.value
+    }));
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    setError('');
+
+    try {
+      const res = await api.post('/join', formData, {
+        headers: { 'Content-Type': 'application/json' },
+        withCredentials: true
+      });
+      alert(`회원가입 완료! 환영합니다, ${res.data.name}님`);
+      navigate('/login');
+    } catch (err) {
+      const msg = err.response?.data?.error || '회원가입에 실패했습니다.';
+      setError(msg);
+    }
+  };
+
+  return (
+    <div className="auth-container">
+      <form className="auth-form" onSubmit={handleSubmit}>
+        <h2>회원가입</h2>
+
+        <input type="text" name="username" placeholder="아이디" value={formData.username} onChange={handleChange} required />
+        <small>※ 영문 + 숫자 조합, 4~20자</small>
+
+        <input type="password" name="password" placeholder="비밀번호" value={formData.password} onChange={handleChange} required />
+        <small>※ 영문, 숫자, 특수문자 포함 8자 이상</small>
+
+        <input type="password" name="passwordCheck" placeholder="비밀번호 확인" value={formData.passwordCheck} onChange={handleChange} required />
+
+        <input type="email" name="email" placeholder="이메일" value={formData.email} onChange={handleChange} required />
+        <small>※ 이메일 형식</small>
+
+        <input type="text" name="name" placeholder="이름" value={formData.name} onChange={handleChange} required />
+        <small>※ 특수문자 없이 한글/영문만</small>
+
+        <button type="submit" className="auth-button">회원가입</button>
+
+        {error && <p className="error-msg">🚫 {error}</p>}
+      </form>
+    </div>
+  );
+};
+
+export default Join;
