+++ src/style/AuthForm.css
@@ -1,0 +1,44 @@
+.auth-container {
+    max-width: 400px;
+    margin: auto;
+    padding: 2rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .auth-form {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 0.8rem;
+  }
+  
+  .auth-form input {
+    padding: 0.8rem;
+    font-size: 1rem;
+  }
+  
+  .auth-form small {
+    color: gray;
+    font-size: 0.8rem;
+    margin-top: -6px;
+    margin-bottom: 8px;
+  }
+  
+  .auth-button {
+    padding: 0.9rem;
+    font-size: 1rem;
+    background-color: #333;
+    color: white;
+    border: none;
+    cursor: pointer;
+    margin-top: 1rem;
+  }
+  
+  .error-msg {
+    color: red;
+    margin-top: 0.5rem;
+    font-size: 0.9rem;
+  }
+
